# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Earth to Date API",
...
    "version": "3.1.0"
  },
  "paths": {
...
    "/pollution/{view}/{zoom}/{tile_x}/{tile_y}": {
      "get": {
        "summary": "Get Pollution Overlay",
        "description": "Low-zoom transparent RGBA pollution OVERLAY tiles (S5P global store).\nSeparate from /v2 (which stays S2 z12-21). Continental scale: zoom 1-9.",
        "operationId": "get_pollution_overlay_pollution__view___zoom___tile_x___tile_y__get",
        "parameters": [
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View"
            }
          },
          {
            "name": "zoom",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Zoom"
            }
          },
          {
            "name": "tile_x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tile X"
            }
          },
          {
            "name": "tile_y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tile Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dem_at_lat_lon": {
      "get": {
        "summary": "Dem At Lat Lon",
        "description": "Get DEM elevation at a specific lat/lon point.\nReturns elevation from all 5 free DEM sources plus the fused result.",
        "operationId": "dem_at_lat_lon_dem_at_lat_lon_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude",
              "title": "Lat"
            },
            "description": "Latitude"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude",
              "title": "Lon"
            },
            "description": "Longitude"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/download_aoi/{bbox_lat_lon}/{date_str}/{days_back}/{max_clouds}/{view}": {
      "get": {
        "summary": "Download Aoi V2",
        "description": "Download Area of Interest - V2 API format\n\nArgs:\n    bbox_lat_lon: Bounding box as \"west,south,east,north\"\n    date_str: Target date in YYYY-MM-DD format\n    days_back: Days to search backwards\n    max_clouds: Maximum cloud coverage percentage\n    view: Product view type",
        "operationId": "download_aoi_v2_v2_download_aoi__bbox_lat_lon___date_str___days_back___max_clouds___view__get",
        "parameters": [
          {
            "name": "bbox_lat_lon",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bbox Lat Lon"
            }
          },
          {
            "name": "date_str",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date Str"
            }
          },
          {
            "name": "days_back",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Days Back"
            }
          },
          {
            "name": "max_clouds",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Max Clouds"
            }
          },
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View"
            }
          },
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Months"
            }
          },
          {
            "name": "prior_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prior Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dates/{lat_lon}/{view}/{date_str}/{days_back}/{max_clouds}": {
      "get": {
        "summary": "Get Available Dates",
        "description": "Get list of available satellite imagery dates for a specific location.\n\nArgs:\n    lat_lon: Location as 'latitude,longitude'\n    view: Product name (e.g., s2_tci, sr_tci)\n    date_str: End date in YYYY-MM-DD format\n    days_back: Days to search back from end date\n    max_clouds: Maximum cloud coverage percentage (0-100)\n    source: Data source - \"s2\" for Sentinel-2 (default) or \"ps\" for PlanetScope\n\nReturns:\n    JSON array of available dates",
        "operationId": "get_available_dates_dates__lat_lon___view___date_str___days_back___max_clouds__get",
        "parameters": [
          {
            "name": "lat_lon",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lat Lon"
            }
          },
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View"
            }
          },
          {
            "name": "date_str",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date Str"
            }
          },
          {
            "name": "days_back",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Days Back"
            }
          },
          {
            "name": "max_clouds",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Max Clouds"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "s2",
              "title": "Source"
            }
          },
          {
            "name": "planet_apikey",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planet Apikey"
            }
          },
          {
            "name": "include_safe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Safe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}